Added some explicit conversion to double precision
When the result of the operation is expected to use double precision, this
patchs adds an explicit conversion to that type to tell the compiler that
this is what we want, and not an unexpected side effect.

+++ WINGs/userdefaults.c
@@ -591,7 +591,7 @@ void WMSetUDFloatForKey(WMUserDefaults * database, float value, const char *defa
 	WMPropList *object;
 	char buffer[128];
 
-	sprintf(buffer, "%f", value);
+	sprintf(buffer, "%f", (double)value);
 	object = WMCreatePLString(buffer);
 
 	WMSetUDObjectForKey(database, object, defaultName);
