+++ WPrefs.app/MouseSettings.c
@@ -301,7 +301,7 @@ static void showData(_Panel * panel)
 	WMSetSliderValue(panel->speedS, (accel - 0.25F) / 0.25F);
 
 	panel->acceleration = accel;
-	sprintf(buffer, "%.2f", accel);
+	sprintf(buffer, "%.2f", (double)accel);
 	WMSetTextFieldText(panel->acceT, buffer);
 
 	 /**/ b = GetIntegerForKey("DoubleClickTime");
